requirements.txt
behave==1.2.6
blindspin==2.0.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
crayons==0.3.0
docker==4.2.0
idna==2.9
parse==1.15.0
parse-type==0.5.2
pika==1.1.0
PyHamcrest==2.0.2
requests==2.23.0
six==1.14.0
testcontainers==2.6.0
urllib3==1.25.8
websocket-client==0.57.0
wrapt==1.12.1
